requirements.txt
gunicorn==20.0.4
Flask==1.1.2
requests==2.23.0
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.3
Pillow==6.2.2
lxml==4.3.3
seaborn==0.11.1
scikit-image
werkzeug==2.0.1
numpy
matplotlib
scipy
